Implemented Somfy RTS / Simu Hz send routines.
This code can be used to enable controlling of
Somfy RTS / Simu Hz blinds and other devices
via FHEM.
The code works the same way as Intertechno,
it switches to the desired frequency, sends the command
and switches back.

It supports changing the interval of a single bit,
and changing the repetition count.
This might help when the receiver is too far away
for reliable transmission.

Thomas Dankert <post@thomyd.de>


git-svn-id: svn://svn.code.sf.net/p/culfw/code/trunk@420 bfc058f7-7f6e-431f-8662-3fa65896f8eb